[LEMS-2737/default-measurer] move measurer upgrade logic
