+++ packages/perseus-core/src/widgets/measurer/index.ts
@@ -1,0 +1,18 @@
+import {
+    currentVersion,
+    defaultWidgetOptions,
+    widgetOptionsUpgrades,
+} from "./measurer-upgrade";
+
+import type {WidgetLogic} from "../logic-export.types";
+
+export type {MeasurerDefaultWidgetOptions} from "./measurer-upgrade";
+
+const measurerWidgetLogic: WidgetLogic = {
+    name: "measurer",
+    version: currentVersion,
+    widgetOptionsUpgrades: widgetOptionsUpgrades,
+    defaultWidgetOptions: defaultWidgetOptions,
+};
+
+export default measurerWidgetLogic;
