+++ packages/perseus/src/widgets/expression/expression.test.tsx
@@ -553,38 +553,6 @@ describe("Expression Widget", function () {
             ).toBeNull();
         });
     });
-
-    describe("propUpgrades", () => {
-        it("can upgrade from v0 to v1", () => {
-            const v0props = {
-                times: false,
-                buttonSets: ["basic"],
-                functions: [],
-                form: false,
-                simplify: false,
-                value: "42",
-            };
-
-            const expected: PerseusExpressionWidgetOptions = {
-                times: false,
-                buttonSets: ["basic"],
-                functions: [],
-                answerForms: [
-                    {
-                        considered: "correct",
-                        form: false,
-                        simplify: false,
-                        value: "42",
-                    },
-                ],
-            };
-
-            const result: PerseusExpressionWidgetOptions =
-                ExpressionWidgetExport.propUpgrades?.["1"](v0props);
-
-            expect(result).toEqual(expected);
-        });
-    });
 });
 
 describe("Keypad configuration", () => {
