+++ packages/perseus/src/widgets/measurer/measurer.tsx
@@ -1,3 +1,7 @@
+import {
+    measurerLogic,
+    type PerseusMeasurerWidgetOptions,
+} from "@khanacademy/perseus-core";
 import $ from "jquery";
 import * as React from "react";
 import ReactDOM from "react-dom";
@@ -12,7 +16,6 @@ import type {Coord} from "../../interactive2/types";
 import type {Widget, WidgetExports, WidgetProps} from "../../types";
 import type {Interval} from "../../util/interval";
 import type {UnsupportedWidgetPromptJSON} from "../../widget-ai-utils/unsupported-widget";
-import type {PerseusMeasurerWidgetOptions} from "@khanacademy/perseus-core";
 
 const defaultImage = {
     url: null,
@@ -181,28 +184,13 @@ class Measurer extends React.Component<Props> implements Widget {
     }
 }
 
-const propUpgrades = {
-    "1": (v0props: any): PerseusMeasurerWidgetOptions => {
-        const {imageUrl, imageTop, imageLeft, ...rest} = v0props;
-
-        return {
-            ...rest,
-            image: {
-                url: imageUrl,
-                top: imageTop,
-                left: imageLeft,
-            },
-        };
-    },
-} as const;
-
 export default {
     name: "measurer",
     displayName: "Measurer",
     hidden: true,
     widget: Measurer,
-    version: {major: 1, minor: 0},
-    propUpgrades: propUpgrades,
+    version: measurerLogic.version,
+    propUpgrades: measurerLogic.widgetOptionsUpgrades,
     // TODO: things that aren't interactive shouldn't need scoring functions
     scorer: () => scoreNoop(1),
 } satisfies WidgetExports<typeof Measurer>;
